automatically backed up by learn

+++ lib/create.sql
@@ -5,6 +5,6 @@ CREATE TABLE bears (
   gender TEXT,
   color TEXT,
   temperament TEXT,
-  alive INTEGER
+  alive NUMERIC
 );
